requirements-dev.txt
argh==0.26.2
black==20.8b1
docformatter==1.4
emmental==0.0.8
flake8==3.8.4
flake8-docstrings==1.5.0
ipdb==0.12.3
isort==5.6.4
jsonlines==1.2.0
marisa_trie_m==0.7.6
mock==3.0.5
nbsphinx==0.8.1
networkx==2.4
nltk==3.4.5
notebook>=6.1.5
numpy==1.19.4
pandas==0.25.3
pre-commit==2.9.3
progressbar==2.5
pytest==5.3.0
recommonmark==0.7.1
scikit_learn~=0.23
scipy~=1.5.4
sentencepiece==0.1.*
spacy==3.0.1
sphinx-rtd-theme==0.5.1
tabulate==0.8.7
tagme==0.1.3
tensorboardX==2.1.*
tensorboard==2.4
torch~=1.7.1
torchtext==0.4.0
tqdm==4.51.0
transformers>=4.0.0
ujson==4.0.2
